+++ app/src/main/AndroidManifest.xml
@@ -15,10 +15,10 @@
         android:theme="@style/Theme.Mumuk"
         tools:targetApi="31">
 
-        <activity android:name=".Signup.SignupActivity" android:exported="false" />
+        <activity android:name=".ui.signup.SignupActivity" android:exported="false" />
         <activity android:name=".ui.MainActivity" android:exported="false" />
-        <activity android:name=".Login.LoginIntroActivity" android:exported="false" />
-        <activity android:name=".Login.LoginActivity" android:exported="false" />
+        <activity android:name=".ui.login.LoginIntroActivity" android:exported="false" />
+        <activity android:name=".ui.login.LoginActivity" android:exported="false" />
         <activity
             android:name=".IntroActivity"
             android:exported="true">
