+++ app/src/main/java/com/example/mumuk/IntroActivity.kt
@@ -8,7 +8,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import com.example.mumuk.Login.LoginIntroActivity
+import com.example.mumuk.ui.login.LoginIntroActivity
 
 class IntroActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
