--- app/src/main/java/com/example/mumuk/Login/LoginIntroActivity.kt
@@ -1,52 +1,0 @@
-package com.example.mumuk.Login
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Button
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import com.example.mumuk.MainActivity
-import com.example.mumuk.R
-import com.example.mumuk.Signup.SignupActivity
-
-class LoginIntroActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContentView(R.layout.activity_login_intro)
-
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-            insets
-        }
-
-
-        findViewById<Button>(R.id.btn_signup).setOnClickListener {
-            startActivity(Intent(this, SignupActivity::class.java))
-            finish()
-        }
-
-        findViewById<Button>(R.id.btn_login).setOnClickListener {
-            startActivity(Intent(this, LoginActivity::class.java))
-            finish()
-        }
-
-        findViewById<Button>(R.id.btn_login_naver).setOnClickListener {
-            startActivity(Intent(this, MainActivity::class.java))
-            finish()
-        }
-
-        findViewById<Button>(R.id.btn_login_kakao).setOnClickListener {
-            startActivity(Intent(this, MainActivity::class.java))
-            finish()
-        }
-
-        findViewById<Button>(R.id.btn_login_samsung).setOnClickListener {
-            startActivity(Intent(this, MainActivity::class.java))
-            finish()
-        }
-    }
-}
