+++ app/src/main/java/com/example/mumuk/ui/login/LoginIntroActivity.kt
@@ -1,0 +1,71 @@
+package com.example.mumuk.ui.login
+
+import android.content.Intent
+import android.os.Bundle
+import android.view.View
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import androidx.fragment.app.commit
+import com.example.mumuk.R
+import com.example.mumuk.databinding.ActivityLoginIntroBinding
+import com.example.mumuk.ui.MainActivity
+import com.example.mumuk.ui.signup.SignupActivity
+
+class LoginIntroActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityLoginIntroBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+
+        binding = ActivityLoginIntroBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        ViewCompat.setOnApplyWindowInsetsListener(binding.main) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }
+
+        binding.btnSignup.setOnClickListener {
+            startActivity(Intent(this, SignupActivity::class.java))
+            finish()
+        }
+
+        binding.btnLogin.setOnClickListener {
+            startActivity(Intent(this, LoginActivity::class.java))
+            finish()
+        }
+
+        binding.btnLoginNaver.setOnClickListener {
+            startActivity(Intent(this, MainActivity::class.java))
+            finish()
+        }
+
+        binding.btnLoginKakao.setOnClickListener {
+            startActivity(Intent(this, MainActivity::class.java))
+            finish()
+        }
+
+        binding.btnLoginSamsung.setOnClickListener {
+            startActivity(Intent(this, MainActivity::class.java))
+            finish()
+        }
+
+        binding.tvFindAccount.setOnClickListener {
+            binding.loginIntroLayout.visibility = View.GONE
+            binding.loginIntroFragmentContainer.visibility = View.VISIBLE
+
+            supportFragmentManager.commit {
+                setReorderingAllowed(true)
+                replace(R.id.login_intro_fragment_container, FindIdFragment())
+                addToBackStack(null)
+            }
+        }
+
+
+    }
+}
