+++ app/src/main/res/drawable/bg_mypage_rounded.xml
@@ -1,0 +1,6 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="@android:color/white"/>
+    <stroke android:width="1dp" android:color="#A2A2A2"/>
+    <corners android:radius="12dp"/>
+</shape>
