+++ app/src/main/res/drawable/btn_cancel.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="120dp"
+    android:height="50dp"
+    android:viewportWidth="120"
+    android:viewportHeight="50">
+  <path
+      android:pathData="M20,0L100,0A20,20 0,0 1,120 20L120,30A20,20 0,0 1,100 50L20,50A20,20 0,0 1,0 30L0,20A20,20 0,0 1,20 0z"
+      android:fillColor="#F3F3F3"/>
+  <path
+      android:pathData="M47.74,18.62H49.1V18.88C49.1,19.76 48.89,20.57 48.48,21.32C48.07,22.05 47.49,22.68 46.74,23.2C45.99,23.72 45.11,24.11 44.1,24.36L43.4,23.06C44.27,22.85 45.03,22.54 45.68,22.14C46.33,21.73 46.84,21.24 47.2,20.68C47.56,20.12 47.74,19.52 47.74,18.88V18.62ZM48.04,18.62H49.4V18.88C49.4,19.48 49.58,20.05 49.94,20.58C50.3,21.1 50.8,21.55 51.44,21.94C52.09,22.33 52.85,22.62 53.7,22.82L53.02,24.12C52.02,23.87 51.15,23.49 50.4,23C49.65,22.49 49.07,21.89 48.66,21.2C48.25,20.49 48.04,19.72 48.04,18.88V18.62ZM43.78,17.52H53.36V18.82H43.78V17.52ZM47.74,15.5H49.4V18.4H47.74V15.5ZM47.78,25.82H49.44V33.02H47.78V25.82ZM55.78,15.48H57.42V33.56H55.78V15.48ZM42.78,27.02L42.56,25.62C43.71,25.62 44.95,25.61 46.3,25.6C47.65,25.57 49.02,25.52 50.42,25.44C51.82,25.36 53.17,25.23 54.48,25.06L54.6,26.28C53.25,26.51 51.88,26.67 50.48,26.78C49.09,26.89 47.74,26.95 46.42,26.98C45.11,27.01 43.9,27.02 42.78,27.02ZM61,29.82H77.4V31.22H61V29.82ZM68.3,25.44H69.94V30.28H68.3V25.44ZM68.24,16.66H69.7V18.06C69.7,18.85 69.56,19.59 69.28,20.28C69.01,20.96 68.64,21.6 68.16,22.2C67.68,22.79 67.13,23.31 66.5,23.78C65.89,24.25 65.23,24.64 64.52,24.96C63.81,25.28 63.09,25.51 62.36,25.66L61.64,24.28C62.28,24.17 62.91,23.99 63.54,23.72C64.18,23.44 64.78,23.11 65.34,22.72C65.9,22.32 66.4,21.88 66.84,21.4C67.28,20.91 67.62,20.38 67.86,19.82C68.11,19.25 68.24,18.66 68.24,18.06V16.66ZM68.54,16.66H69.98V18.06C69.98,18.67 70.11,19.26 70.36,19.82C70.61,20.38 70.96,20.91 71.4,21.4C71.84,21.88 72.34,22.32 72.9,22.72C73.47,23.12 74.07,23.45 74.7,23.72C75.34,23.99 75.98,24.17 76.62,24.28L75.9,25.66C75.17,25.51 74.45,25.28 73.74,24.96C73.03,24.64 72.37,24.25 71.74,23.78C71.11,23.31 70.56,22.79 70.08,22.2C69.6,21.6 69.22,20.95 68.94,20.26C68.67,19.57 68.54,18.83 68.54,18.06V16.66Z"
+      android:fillColor="#A2A2A2"/>
+</vector>
