+++ app/src/main/res/drawable/btn_done.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="120dp"
+    android:height="50dp"
+    android:viewportWidth="120"
+    android:viewportHeight="50">
+  <path
+      android:pathData="M20,0L100,0A20,20 0,0 1,120 20L120,30A20,20 0,0 1,100 50L20,50A20,20 0,0 1,0 30L0,20A20,20 0,0 1,20 0z"
+      android:fillColor="#007251"/>
+  <path
+      android:pathData="M47.28,22.98H48.94V25.72H47.28V22.98ZM48.06,16.3C48.93,16.3 49.69,16.45 50.36,16.76C51.04,17.05 51.57,17.47 51.94,18C52.33,18.53 52.52,19.15 52.52,19.86C52.52,20.57 52.33,21.19 51.94,21.74C51.57,22.27 51.04,22.69 50.36,23C49.69,23.29 48.93,23.44 48.06,23.44C47.19,23.44 46.42,23.29 45.74,23C45.07,22.69 44.55,22.27 44.16,21.74C43.79,21.19 43.6,20.57 43.6,19.86C43.6,19.15 43.79,18.53 44.16,18C44.55,17.47 45.07,17.05 45.74,16.76C46.42,16.45 47.19,16.3 48.06,16.3ZM48.06,17.58C47.5,17.58 47,17.67 46.56,17.86C46.12,18.05 45.77,18.31 45.52,18.66C45.28,19.01 45.16,19.41 45.16,19.86C45.16,20.31 45.28,20.71 45.52,21.06C45.77,21.41 46.12,21.68 46.56,21.88C47,22.07 47.5,22.16 48.06,22.16C48.63,22.16 49.13,22.07 49.56,21.88C50,21.68 50.34,21.41 50.58,21.06C50.83,20.71 50.96,20.31 50.96,19.86C50.96,19.41 50.83,19.01 50.58,18.66C50.34,18.31 50,18.05 49.56,17.86C49.13,17.67 48.63,17.58 48.06,17.58ZM54.96,15.48H56.62V29.32H54.96V15.48ZM55.92,21.58H59.26V22.98H55.92V21.58ZM45.18,31.78H57.26V33.14H45.18V31.78ZM45.18,28.14H46.84V32.38H45.18V28.14ZM42.78,26.64L42.58,25.26C43.63,25.26 44.79,25.25 46.06,25.24C47.34,25.21 48.65,25.16 49.98,25.08C51.31,24.99 52.59,24.85 53.82,24.66L53.94,25.88C52.69,26.11 51.39,26.27 50.06,26.38C48.74,26.49 47.45,26.56 46.2,26.6C44.96,26.63 43.82,26.64 42.78,26.64ZM65.58,26.06H67.2V30.5H65.58V26.06ZM71.36,26.04H72.98V30.48H71.36V26.04ZM61,29.94H77.4V31.32H61V29.94ZM63,16.8H75.36V22.28H64.68V25.9H63.04V20.94H73.72V18.16H63V16.8ZM63.04,25.18H75.78V26.54H63.04V25.18Z"
+      android:fillColor="#ffffff"/>
+</vector>
