+++ app/src/main/res/drawable/btn_mypage_select.xml
@@ -1,0 +1,8 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="@android:color/transparent" />
+    <corners android:radius="12dp" />
+    <stroke
+        android:width="1dp"
+        android:color="#F3F3F3" />
+</shape>
