+++ app/src/main/res/drawable/ic_help.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M16,1H8V5H16V1Z"
+      android:fillColor="#000000"/>
+  <path
+      android:pathData="M3,3H6V7H18V3H21V23H3V3Z"
+      android:fillColor="#000000"/>
+</vector>
