+++ app/src/main/res/drawable/ic_logout.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M5,21C4.45,21 3.979,20.804 3.588,20.413C3.197,20.022 3.001,19.551 3,19V5C3,4.45 3.196,3.979 3.588,3.588C3.98,3.197 4.451,3.001 5,3H12V5H5V19H12V21H5ZM16,17L14.625,15.55L17.175,13H9V11H17.175L14.625,8.45L16,7L21,12L16,17Z"
+      android:fillColor="#000000"/>
+</vector>
