+++ app/src/main/res/drawable/ic_profile_pen.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M14.757,2.621C15.635,1.743 16.826,1.25 18.068,1.25C19.31,1.25 20.501,1.743 21.379,2.621C22.257,3.5 22.75,4.691 22.75,5.932C22.75,7.174 22.257,8.365 21.379,9.243L11.893,18.729C11.351,19.271 11.033,19.59 10.677,19.867C10.258,20.194 9.808,20.473 9.327,20.701C8.921,20.895 8.493,21.038 7.767,21.279L4.435,22.389L3.633,22.657C3.314,22.764 2.972,22.779 2.644,22.702C2.317,22.625 2.018,22.458 1.78,22.221C1.542,21.983 1.375,21.683 1.298,21.356C1.221,21.029 1.236,20.686 1.343,20.368L2.721,16.235C2.963,15.507 3.106,15.08 3.299,14.672C3.528,14.193 3.807,13.743 4.134,13.323C4.41,12.969 4.729,12.649 5.271,12.108L14.757,2.621ZM4.4,20.822L7.241,19.874C8.032,19.61 8.368,19.497 8.681,19.347C9.062,19.165 9.42,18.944 9.754,18.684C10.027,18.471 10.279,18.222 10.869,17.632L18.439,10.061C17.401,9.694 16.459,9.098 15.682,8.317C14.902,7.54 14.307,6.598 13.94,5.56L6.37,13.13C5.78,13.719 5.53,13.97 5.317,14.245C5.057,14.578 4.836,14.936 4.654,15.318C4.505,15.63 4.392,15.967 4.128,16.757L3.18,19.601L4.4,20.822ZM15.155,4.343C15.19,4.518 15.247,4.756 15.344,5.033C15.636,5.871 16.115,6.63 16.744,7.255C17.369,7.884 18.128,8.363 18.965,8.655C19.243,8.752 19.481,8.809 19.656,8.844L20.318,8.182C20.911,7.585 21.243,6.777 21.242,5.935C21.24,5.093 20.905,4.286 20.31,3.69C19.715,3.095 18.908,2.76 18.066,2.758C17.224,2.757 16.415,3.089 15.818,3.682L15.155,4.343Z"
+      android:fillColor="#000000"
+      android:fillType="evenOdd"/>
+</vector>
