+++ app/src/main/res/drawable/ic_settings.xml
@@ -1,9 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="388dp"
+    android:width="24dp"
     android:height="24dp"
-    android:viewportWidth="388"
+    android:viewportWidth="24"
     android:viewportHeight="24">
   <path
-      android:pathData="M9.25,22L8.85,18.8C8.63,18.72 8.43,18.62 8.24,18.5C8.05,18.38 7.86,18.26 7.67,18.13L4.7,19.38L1.95,14.63L4.52,12.68C4.51,12.56 4.5,12.45 4.5,12.34V11.66C4.5,11.55 4.51,11.44 4.52,11.32L1.95,9.38L4.7,4.63L7.67,5.88C7.86,5.74 8.05,5.62 8.25,5.5C8.45,5.38 8.65,5.28 8.85,5.2L9.25,2H14.75L15.15,5.2C15.37,5.28 15.57,5.38 15.76,5.5C15.95,5.62 16.14,5.74 16.33,5.88L19.3,4.63L22.05,9.38L19.48,11.32C19.49,11.44 19.5,11.55 19.5,11.66V12.34C19.5,12.45 19.48,12.56 19.45,12.68L22.02,14.63L19.27,19.38L16.33,18.13C16.14,18.26 15.95,18.38 15.75,18.5C15.55,18.62 15.35,18.72 15.15,18.8L14.75,22H9.25ZM11,20H12.98L13.32,17.35C13.84,17.22 14.32,17.02 14.76,16.76C15.2,16.5 15.61,16.19 15.98,15.82L18.45,16.85L19.42,15.15L17.27,13.52C17.36,13.29 17.42,13.05 17.45,12.79C17.48,12.53 17.5,12.27 17.5,12C17.5,11.73 17.48,11.47 17.45,11.21C17.42,10.96 17.36,10.71 17.27,10.48L19.42,8.85L18.45,7.15L15.98,8.2C15.61,7.82 15.2,7.5 14.76,7.24C14.32,6.98 13.84,6.78 13.32,6.65L13,4H11.02L10.68,6.65C10.16,6.78 9.68,6.98 9.24,7.24C8.8,7.5 8.39,7.81 8.02,8.18L5.55,7.15L4.57,8.85L6.72,10.45C6.64,10.7 6.58,10.95 6.55,11.2C6.52,11.45 6.5,11.72 6.5,12C6.5,12.27 6.52,12.52 6.55,12.77C6.58,13.02 6.64,13.27 6.72,13.52L4.57,15.15L5.55,16.85L8.02,15.8C8.39,16.18 8.8,16.5 9.24,16.76C9.68,17.02 10.16,17.22 10.68,17.35L11,20ZM12.05,15.5C13.02,15.5 13.84,15.16 14.52,14.48C15.21,13.79 15.55,12.97 15.55,12C15.55,11.03 15.21,10.21 14.52,9.52C13.84,8.84 13.02,8.5 12.05,8.5C11.07,8.5 10.24,8.84 9.56,9.52C8.89,10.21 8.55,11.03 8.55,12C8.55,12.97 8.89,13.79 9.56,14.48C10.24,15.16 11.07,15.5 12.05,15.5Z"
+      android:pathData="M9.25,22L8.85,18.8C8.633,18.717 8.429,18.617 8.238,18.5C8.047,18.383 7.859,18.258 7.675,18.125L4.7,19.375L1.95,14.625L4.525,12.675C4.508,12.558 4.5,12.446 4.5,12.338V11.663C4.5,11.554 4.508,11.442 4.525,11.325L1.95,9.375L4.7,4.625L7.675,5.875C7.858,5.742 8.05,5.617 8.25,5.5C8.45,5.383 8.65,5.283 8.85,5.2L9.25,2H14.75L15.15,5.2C15.367,5.283 15.571,5.383 15.763,5.5C15.955,5.617 16.142,5.742 16.325,5.875L19.3,4.625L22.05,9.375L19.475,11.325C19.492,11.442 19.5,11.554 19.5,11.663V12.337C19.5,12.446 19.483,12.558 19.45,12.675L22.025,14.625L19.275,19.375L16.325,18.125C16.142,18.258 15.95,18.383 15.75,18.5C15.55,18.617 15.35,18.717 15.15,18.8L14.75,22H9.25ZM11,20H12.975L13.325,17.35C13.842,17.217 14.321,17.021 14.763,16.763C15.205,16.505 15.609,16.192 15.975,15.825L18.45,16.85L19.425,15.15L17.275,13.525C17.358,13.292 17.417,13.046 17.45,12.788C17.483,12.53 17.5,12.267 17.5,12C17.5,11.733 17.483,11.47 17.45,11.213C17.417,10.956 17.358,10.71 17.275,10.475L19.425,8.85L18.45,7.15L15.975,8.2C15.608,7.817 15.204,7.496 14.763,7.238C14.322,6.98 13.842,6.784 13.325,6.65L13,4H11.025L10.675,6.65C10.158,6.783 9.679,6.979 9.238,7.238C8.797,7.497 8.392,7.809 8.025,8.175L5.55,7.15L4.575,8.85L6.725,10.45C6.642,10.7 6.583,10.95 6.55,11.2C6.517,11.45 6.5,11.717 6.5,12C6.5,12.267 6.517,12.525 6.55,12.775C6.583,13.025 6.642,13.275 6.725,13.525L4.575,15.15L5.55,16.85L8.025,15.8C8.392,16.183 8.796,16.504 9.238,16.763C9.68,17.022 10.159,17.217 10.675,17.35L11,20ZM12.05,15.5C13.017,15.5 13.842,15.158 14.525,14.475C15.208,13.792 15.55,12.967 15.55,12C15.55,11.033 15.208,10.208 14.525,9.525C13.842,8.842 13.017,8.5 12.05,8.5C11.067,8.5 10.237,8.842 9.562,9.525C8.887,10.208 8.549,11.033 8.55,12C8.551,12.967 8.888,13.792 9.563,14.475C10.238,15.158 11.067,15.5 12.05,15.5Z"
       android:fillColor="#000000"/>
 </vector>
