+++ app/src/main/res/drawable/ic_version.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M12,1.5C9.923,1.5 7.893,2.116 6.167,3.27C4.44,4.423 3.094,6.063 2.299,7.982C1.505,9.9 1.297,12.012 1.702,14.048C2.107,16.085 3.107,17.956 4.575,19.425C6.044,20.893 7.915,21.893 9.952,22.298C11.988,22.703 14.1,22.496 16.018,21.701C17.937,20.906 19.577,19.56 20.73,17.833C21.884,16.107 22.5,14.077 22.5,12C22.5,9.215 21.394,6.545 19.425,4.575C17.455,2.606 14.785,1.5 12,1.5ZM12,6C12.222,6 12.44,6.066 12.625,6.19C12.81,6.313 12.954,6.489 13.039,6.694C13.125,6.9 13.147,7.126 13.103,7.344C13.06,7.563 12.953,7.763 12.795,7.92C12.638,8.078 12.438,8.185 12.219,8.228C12.001,8.272 11.775,8.25 11.569,8.164C11.364,8.079 11.188,7.935 11.065,7.75C10.941,7.565 10.875,7.347 10.875,7.125C10.875,6.827 10.993,6.54 11.205,6.33C11.415,6.119 11.702,6 12,6ZM15,18.094H9V16.406H11.156V12.094H9.75V10.406H12.844V16.406H15V18.094Z"
+      android:fillColor="#000000"/>
+</vector>
