+++ app/src/main/res/layout/dialog_logout.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="200dp"
+    android:layout_height="100dp"
+    android:orientation="vertical"
+    android:background="@drawable/btn_dialog_bg"
+    android:gravity="top|center_horizontal"
+    android:paddingStart="12dp"
+    android:paddingEnd="12dp">
+
+    <TextView
+        android:id="@+id/tv_dialog_message"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="15dp"
+        android:text="로그아웃 되었습니다."
+        android:textSize="12sp"
+        android:textStyle="bold"
+        android:textColor="@color/black"
+        android:gravity="center"
+         />
+
+    <TextView
+        android:id="@+id/tv_dialog_submessage"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="2dp"
+        android:text="다시 로그인하세요"
+        android:textSize="12sp"
+        android:textStyle="bold"
+        android:textColor="#A2A2A2"
+        android:gravity="center"
+         />
+
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="1dp"
+        android:layout_marginTop="15dp"
+        android:layout_marginBottom="0dp"
+        android:background="#F3F3F3" />
+
+    <TextView
+        android:id="@+id/btn_dialog_ok"
+        android:layout_width="match_parent"
+        android:layout_height="30dp"
+        android:layout_marginTop="3dp"
+        android:gravity="center"
+        android:text="로그인"
+        android:textColor="#306AF2"
+        android:textStyle="bold"
+        android:textSize="12sp"
+        android:clickable="true"
+        android:focusable="true"/>
+</LinearLayout>
